Start on a new version in Rust
The node-based version has proved to be annoying to maintain. I've been
thinking about doing a rewrite in Rust for quite some time now, but
finally got around to it.

This is still just experimentation on whether I can create a thumbnail
and get keywords without large, external libraries like ImageMagick. It
seems like I can. The image-crate is very good for making the thumbnail.

The keywords are a bit trickier. There doesn't seem be any rust implementation for
extracting IPTC metadata from a JPEG. I forked jpeg-decoder to have it
expose Photoshop IRB data while it's decoding the image. Here, I parse
the IRBs just enough to find the IPTC data and then parse those just
enough to find the keywords.

The next steps would be connection to AWS S3 and DynamoDB for storing
the thumbnails and the metadata, including the keywords.